Do not kill redis

+++ scripts/main.py
@@ -90,7 +90,8 @@ def run(type, server_conf, memtier_conf, output_dir):
     #     print("[Main] CPU Average: " + str(cpu_average))
 
     # Clean up
-    server.kill()
+    if type != 'redis':
+        server.kill()
 
 
 def parse_config(path):
